# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation DADALI-CAMPERNIK API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/mobile/dashboard": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get dashboard data",
                "description": "Mengambil data dashboard untuk user yang sedang login, termasuk informasi profile dan status proyek.",
                "operationId": "73eb4217a0b8b3e442aa9b48349ebad3",
                "responses": {
                    "200": {
                        "description": "Dashboard data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dashboard data"
                                        },
                                        "data": {
                                            "properties": {
                                                "profile": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Abdul"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "abdul@example.com"
                                                        },
                                                        "phone": {
                                                            "type": "string",
                                                            "example": "08123456789"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "project_status": {
                                                    "properties": {
                                                        "total_proyek": {
                                                            "type": "integer",
                                                            "example": 25
                                                        },
                                                        "total_on_progress": {
                                                            "type": "integer",
                                                            "example": 10
                                                        },
                                                        "total_selesai": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "total_active": {
                                                            "type": "integer",
                                                            "example": 20
                                                        },
                                                        "total_inactive": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "total_dokumen": {
                                                            "type": "integer",
                                                            "example": 50
                                                        },
                                                        "total_pengawasan": {
                                                            "type": "integer",
                                                            "example": 30
                                                        },
                                                        "total_kendala": {
                                                            "type": "integer",
                                                            "example": 2
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Token tidak valid"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/mobile/laporan/kendalas": {
            "get": {
                "tags": [
                    "LaporanKendala"
                ],
                "summary": "History Laporan Kendala",
                "description": "List history laproan kendala dengan search",
                "operationId": "mobile/kendalas",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Keyword pencarian (judul atau nama proyek)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "cuaca badai"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return list hstory kendala"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/mobile/laporan/kendala": {
            "post": {
                "tags": [
                    "LaporanKendala"
                ],
                "summary": "Laporan khusus kendala atau insidentil",
                "operationId": "db08fc2673dd7c1c8992f4b795d4bb9d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "pekerjaan_id",
                                    "foto_url",
                                    "tipe_masalah",
                                    "judul",
                                    "deskripsi"
                                ],
                                "properties": {
                                    "pekerjaan_id": {
                                        "type": "integer",
                                        "example": 7
                                    },
                                    "tipe_masalah": {
                                        "type": "string",
                                        "enum": [
                                            "teknis",
                                            "administrasi",
                                            "anggaran",
                                            "koordinasi",
                                            "lingkungan",
                                            "regulasi",
                                            "alam",
                                            "lainnya"
                                        ],
                                        "example": "teknis"
                                    },
                                    "judul": {
                                        "type": "string",
                                        "example": "Kualitas material tidak sesuai spesifikasi"
                                    },
                                    "deskripsi": {
                                        "type": "string",
                                        "example": "Bahan yang tidak sesuai dengan perencanaan"
                                    },
                                    "foto_url": {
                                        "description": "Upload foto (jpg, jpeg, png, max 5MB)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "return model laporan_kendala"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "securityScheme": "bearerAuth",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}